requirements.txt
lhotse>=1.31.1
yamlargparse>=1.31.1
gdown>=5.2.0
torchaudio>=2.8.0
datasets>=4.2.0
pandas>=2.3.3
matplotlib>=3.10.7
python-dotenv>=1.1.1
awscli>=1.42.51
boto3>=1.40.51
ego4d>=1.7.3
einops>=0.8.1

# Training pipeline dependencies
accelerate>=1.10.1
dacite>=1.9.2
tensorboard>=2.20.0
wandb>=0.22.2
tqdm>=4.67.1
typing-extensions>=4.15.0
textgrid>=1.6.1